+++ tests/Doctrine/Tests/Mocks/EntityManagerMock.php
@@ -7,8 +7,8 @@ namespace Doctrine\Tests\Mocks;
 use Doctrine\Common\EventManager;
 use Doctrine\ORM\Configuration;
 use Doctrine\ORM\EntityManager;
+use Doctrine\ORM\ORMSetup;
 use Doctrine\ORM\Proxy\ProxyFactory;
-use Doctrine\ORM\Tools\DoctrineSetup;
 use Doctrine\ORM\UnitOfWork;
 
 /**
@@ -61,7 +61,7 @@ class EntityManagerMock extends EntityManager
             $config = new Configuration();
             $config->setProxyDir(__DIR__ . '/../Proxies');
             $config->setProxyNamespace('Doctrine\Tests\Proxies');
-            $config->setMetadataDriverImpl(DoctrineSetup::createDefaultAnnotationDriver());
+            $config->setMetadataDriverImpl(ORMSetup::createDefaultAnnotationDriver());
         }
 
         if ($eventManager === null) {
